[test] MarketChartDtoMapper #254
This commit adds comprehensive tests for the `MarketChartDtoMapper` which is responsible for converting `MarketChartResponse` to `Chart`.

The tests cover various scenarios:

-   Conversion of valid market chart responses.
-   Handling of empty prices.
-   Handling of negative timestamps and prices.
-   Handling of zero values.
-   Handling of large double values.
-   Truncation of decimals in non-integer timestamps.
-   Handling of missing elements in price pairs, throwing `IndexOutOfBoundsException`.

+++ domain/market/src/main/java/ir/composenews/domain/model/Chart.kt
@@ -3,5 +3,5 @@ package ir.composenews.domain.model
 import kotlinx.collections.immutable.PersistentList
 
 data class Chart(
-    val prices: PersistentList<Pair<Int, Double>>,
+    val prices: PersistentList<Pair<Long, Double>>,
 )
